par de errorcillos

+++ Matricula/Matricula/aranceles.cpp
@@ -3,7 +3,7 @@
 float Aranceles::cobroMatricula(Curso *C, Estudiante *E)
 {
 	float resultado = 0;
-	resultado = C->getCantidadCreditos() * costoCredito;
+	resultado = (float) C->getCantidadCreditos() * costoCredito;
 	resultado = resultado + cargosAdministrativos;
 	if (E->getNacionalidad())
 		resultado = resultado + (resultado * recargoExtranjeros);
