+++ Matricula/Matricula/aranceles.h
@@ -9,7 +9,7 @@ class Aranceles
 private:
 	const int costoCredito = 10000;
 	const int cargosAdministrativos = 15000;
-	const int recargoExtranjeros = 0.40;
+	const double recargoExtranjeros = 0.40;
 public:
 	float cobroMatricula(Curso*, Estudiante*);
